requirements.txt
gradio~=5.43.1
tensorflow~=2.20.0
numpy~=2.2.6
pillow~=11.3.0
scikit-learn~=1.7.1